+++ packages/expo-av/src/ExponentAV.web.ts
@@ -92,10 +92,10 @@ function getStatusFromMedia(media?: HTMLMediaElement): AVPlaybackStatus {
   return status;
 }
 
-function setStatusForMedia(
+async function setStatusForMedia(
   media: HTMLMediaElement,
   status: AVPlaybackStatusToSet
-): AVPlaybackStatus {
+): Promise<AVPlaybackStatus> {
   if (status.positionMillis !== undefined) {
     media.currentTime = status.positionMillis / 1000;
   }
@@ -113,9 +113,9 @@ function setStatusForMedia(
   // }
   if (status.shouldPlay !== undefined) {
     if (status.shouldPlay) {
-      media.play();
+      await media.play();
     } else {
-      media.pause();
+      await media.pause();
     }
   }
   if (status.rate !== undefined) {
@@ -203,7 +203,7 @@ export default {
       });
     };
 
-    const status = setStatusForMedia(media, fullInitialStatus);
+    const status = await setStatusForMedia(media, fullInitialStatus);
 
     return [media, status];
   },